Fix all manual_div_ceil clippy warnings (228 instances)
Applied Rust 1.73+ div_ceil() method to replace manual ceiling division
pattern (x + y - 1) / y with x.div_ceil(y) across the entire codebase.

Scope:
- 207 files modified
- 228 manual_div_ceil warnings eliminated
- All instances replaced with idiomatic div_ceil() method

Files affected:
- trustformers-core: GPU kernels, BLAS, export formats, quantization
- trustformers-models: Model implementations (GPT-2, T5, StableLM, etc.)
- trustformers-serve: Resource management, batching, monitoring
- trustformers-training: Parallelism, RLHF trainer
- trustformers-wasm: WebGPU operations, memory management
- trustformers-c: Codegen, performance analytics
- trustformers-optim: Kernel fusion, quantization

Common patterns replaced:
- (n + 15) / 16 → n.div_ceil(16)  (GPU workgroup sizing)
- (size + 255) / 256 → size.div_ceil(256)  (Thread dispatch)
- (total + cols - 1) / cols → total.div_ceil(cols)  (Grid calculations)

Testing:
- Verified flash attention tests still pass (20/20)
- Build succeeds cleanly
- No functional changes, only idiomatic improvements

Benefits:
- More readable code
- Consistent with modern Rust idioms
- Eliminates potential overflow issues in manual calculations
- Clippy-clean codebase for alpha.2 release

+++ trustformers-serve/src/memory_pressure/cleanup/handlers.rs
@@ -488,7 +488,7 @@ impl ModelUnloadingHandler {
         }
 
         // Simulate model unloading time
-        let unload_time = Duration::from_millis(50 + (model_size / (1024 * 1024)) as u64);
+        let unload_time = Duration::from_millis(50 + (model_size / (1024 * 1024)));
         std::thread::sleep(unload_time);
 
         info!(
